Edit ingredientes_recetas

+++ database/migrations/2024_02_26_184429_create_ingredientes_recetas_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->integer('receta_id');
             $table->integer('ingrediente_id');
-            $table->integer('cantidad');
+            $table->decimal('cantidad');
             $table->string('unidad');
             $table->timestamps();
         });
